Refactor reservation unit graphql objects to separate folders
