+++ api/graphql/types/equipment_category/permissions.py
@@ -1,0 +1,16 @@
+from typing import Any
+
+from graphene_permissions.permissions import BasePermission
+
+from common.typing import GQLInfo
+from permissions.helpers import can_manage_equipment_categories
+
+
+class EquipmentCategoryPermission(BasePermission):
+    @classmethod
+    def has_permission(cls, info: GQLInfo) -> bool:
+        return True
+
+    @classmethod
+    def has_mutation_permission(cls, root: Any, info: GQLInfo, input: dict) -> bool:
+        return can_manage_equipment_categories(info.context.user)
