+++ api/graphql/types/purpose/serializers.py
@@ -1,0 +1,17 @@
+from api.graphql.extensions.legacy_helpers import (
+    OldPrimaryKeySerializer,
+    OldPrimaryKeyUpdateSerializer,
+    get_all_translatable_fields,
+)
+from reservation_units.models import Purpose
+
+
+class PurposeCreateSerializer(OldPrimaryKeySerializer):
+    class Meta:
+        model = Purpose
+        fields = get_all_translatable_fields(model)
+
+
+class PurposeUpdateSerializer(OldPrimaryKeyUpdateSerializer, PurposeCreateSerializer):
+    class Meta(PurposeCreateSerializer.Meta):
+        fields = ["pk"] + PurposeCreateSerializer.Meta.fields
