+++ api/graphql/types/reservation_unit_cancellation_rule/permissions.py
@@ -1,0 +1,19 @@
+from typing import Any
+
+from graphene_permissions.permissions import BasePermission
+
+from common.typing import GQLInfo
+
+
+class ReservationUnitCancellationRulePermission(BasePermission):
+    @classmethod
+    def has_permission(cls, info: GQLInfo) -> bool:
+        return info.context.user.is_authenticated
+
+    @classmethod
+    def has_filter_permission(cls, info: GQLInfo) -> bool:
+        return info.context.user.is_authenticated
+
+    @classmethod
+    def has_mutation_permission(cls, root: Any, info: GQLInfo, input: dict) -> bool:
+        return False
