+++ api/graphql/types/reservation_units/mutations.py
@@ -1,145 +1,28 @@
 import graphene
 from auditlog.models import LogEntry
 from django.conf import settings
-from graphene import ClientIDMutation
 from graphene_django.rest_framework.mutation import SerializerMutation
-from graphene_file_upload.scalars import Upload
 from graphql import GraphQLError
-from rest_framework.generics import get_object_or_404
 
-from api.graphql.extensions.legacy_helpers import OldAuthDeleteMutation, OldAuthSerializerMutation
+from api.graphql.extensions.legacy_helpers import OldAuthSerializerMutation
 from api.graphql.types.reservation_units.permissions import (
-    EquipmentCategoryPermission,
-    EquipmentPermission,
-    PurposePermission,
-    ReservationUnitImagePermission,
     ReservationUnitPermission,
 )
 from api.graphql.types.reservation_units.serializers import (
-    EquipmentCategoryCreateSerializer,
-    EquipmentCategoryUpdateSerializer,
-    EquipmentCreateSerializer,
-    EquipmentUpdateSerializer,
-    PurposeCreateSerializer,
-    PurposeUpdateSerializer,
     ReservationUnitCreateSerializer,
-    ReservationUnitImageCreateSerializer,
-    ReservationUnitImageUpdateSerializer,
     ReservationUnitUpdateSerializer,
 )
 from api.graphql.types.reservation_units.types import (
-    EquipmentCategoryType,
-    EquipmentType,
-    PurposeType,
-    ReservationUnitImageType,
     ReservationUnitType,
 )
 from opening_hours.errors import HaukiAPIError, HaukiRequestError
 from opening_hours.utils.hauki_resource_hash_updater import HaukiResourceHashUpdater
 from reservation_units.models import (
-    Equipment,
-    EquipmentCategory,
-    Purpose,
     ReservationUnit,
-    ReservationUnitImage,
 )
 from tilavarauspalvelu.utils import logging
 
 logger = logging.getLogger(__name__)
-
-
-class EquipmentCreateMutation(OldAuthSerializerMutation, SerializerMutation):
-    equipment = graphene.Field(EquipmentType)
-
-    permission_classes = (EquipmentPermission,)
-
-    class Meta:
-        model_operations = ["create"]
-        serializer_class = EquipmentCreateSerializer
-
-
-class EquipmentUpdateMutation(OldAuthSerializerMutation, SerializerMutation):
-    equipment = graphene.Field(EquipmentType)
-
-    permission_classes = (EquipmentPermission,)
-
-    class Meta:
-        model_operations = ["update"]
-        lookup_field = "pk"
-        serializer_class = EquipmentUpdateSerializer
-
-
-class EquipmentDeleteMutation(OldAuthDeleteMutation, ClientIDMutation):
-    permission_classes = (EquipmentPermission,)
-    model = Equipment
-
-    @classmethod
-    def validate(cls, root, info, **input):
-        return None
-
-
-class EquipmentCategoryCreateMutation(OldAuthSerializerMutation, SerializerMutation):
-    equipment_category = graphene.Field(EquipmentCategoryType)
-
-    permission_classes = (EquipmentCategoryPermission,)
-
-    class Meta:
-        model_operations = ["create"]
-        serializer_class = EquipmentCategoryCreateSerializer
-
-
-class EquipmentCategoryUpdateMutation(OldAuthSerializerMutation, SerializerMutation):
-    equipment_category = graphene.Field(EquipmentCategoryType)
-
-    permission_classes = (EquipmentCategoryPermission,)
-
-    class Meta:
-        model_operations = ["update"]
-        lookup_field = "pk"
-        serializer_class = EquipmentCategoryUpdateSerializer
-
-
-class EquipmentCategoryDeleteMutation(OldAuthDeleteMutation, ClientIDMutation):
-    permission_classes = (EquipmentCategoryPermission,)
-    model = EquipmentCategory
-
-    @classmethod
-    def validate(cls, root, info, **input):
-        return None
-
-
-class PurposeCreateMutation(OldAuthSerializerMutation, SerializerMutation):
-    purpose = graphene.Field(PurposeType)
-
-    permission_classes = (PurposePermission,)
-
-    class Meta:
-        model_operations = ["create"]
-
-        serializer_class = PurposeCreateSerializer
-
-    @classmethod
-    def perform_mutate(cls, serializer, info):
-        purpose = serializer.create(serializer.validated_data)
-        return cls(errors=None, purpose=purpose)
-
-
-class PurposeUpdateMutation(OldAuthSerializerMutation, SerializerMutation):
-    purpose = graphene.Field(PurposeType)
-
-    permission_classes = (PurposePermission,)
-
-    class Meta:
-        model_operations = ["update"]
-        lookup_field = "pk"
-        serializer_class = PurposeUpdateSerializer
-
-    @classmethod
-    def perform_mutate(cls, serializer, info):
-        validated_data = serializer.validated_data
-        pk = validated_data.get("pk")
-        purpose = serializer.update(get_object_or_404(Purpose, pk=pk), validated_data)
-        return cls(errors=None, purpose=purpose)
 
 
 class ReservationUnitMutationMixin:
@@ -222,41 +105,3 @@ class ReservationUnitUpdateMutation(ReservationUnitMutationMixin, OldAuthSeriali
         model_operations = ["update"]
         lookup_field = "pk"
         serializer_class = ReservationUnitUpdateSerializer
-
-
-class ReservationUnitImageCreateMutation(OldAuthSerializerMutation, SerializerMutation):
-    reservation_unit_image = graphene.Field(ReservationUnitImageType)
-
-    class Input:
-        image = Upload()
-
-    permission_classes = (ReservationUnitImagePermission,)
-
-    class Meta:
-        model_operations = ["create"]
-        serializer_class = ReservationUnitImageCreateSerializer
-
-    def resolve_reservation_unit_image(self, info):
-        if self.pk:
-            return get_object_or_404(ReservationUnitImage, pk=self.pk)
-        return None
-
-
-class ReservationUnitImageUpdateMutation(OldAuthSerializerMutation, SerializerMutation):
-    reservation_unit_image = graphene.Field(ReservationUnitImageType)
-
-    permission_classes = (ReservationUnitImagePermission,)
-
-    class Meta:
-        model_operations = ["update"]
-        lookup_field = "pk"
-        serializer_class = ReservationUnitImageUpdateSerializer
-
-
-class ReservationUnitImageDeleteMutation(OldAuthDeleteMutation, ClientIDMutation):
-    permission_classes = (ReservationUnitImagePermission,)
-    model = ReservationUnitImage
-
-    @classmethod
-    def validate(cls, root, info, **input):
-        return None
